Extract styled sub components into extra files
