+++ src/components/MappingEvents/ChevronLeft.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+import ChevronLeft from '../ChevronLeft';
+
+export const StyledChevronLeft = styled(ChevronLeft)`
+  margin-left: 0;
+  padding: 15px 30px 15px 0;
+  z-index: 4;
+`;
+
+export default StyledChevronLeft;
